Birth of a new god
