--- Ragdoll/src/Ragdoll/Components/RenderableLayer.cpp
@@ -1,3 +1,0 @@
-#include "ragdollpch.h"
-
-#include "RenderableLayer.h"
