+++ Ragdoll/src/Ragdoll/GLTFLoader.h
@@ -6,7 +6,7 @@ class ForwardRenderer;
 namespace ragdoll {
 	class FileManager;
 	class EntityManager;
-	class TransformSystem;
+	class Scene;
 }
 
 class GLTFLoader {
@@ -14,9 +14,9 @@ class GLTFLoader {
 	ForwardRenderer* Renderer;
 	std::shared_ptr<ragdoll::FileManager> FileManager;
 	std::shared_ptr<ragdoll::EntityManager> EntityManager;
-	std::shared_ptr<ragdoll::TransformSystem> TransformLayer;
+	std::shared_ptr<ragdoll::Scene> Scene;
 public:
-	void Init(std::filesystem::path root, ForwardRenderer* renderer, std::shared_ptr<ragdoll::FileManager> fm, std::shared_ptr<ragdoll::EntityManager> em, std::shared_ptr<ragdoll::TransformSystem> tl);
+	void Init(std::filesystem::path root, ForwardRenderer* renderer, std::shared_ptr<ragdoll::FileManager> fm, std::shared_ptr<ragdoll::EntityManager> em, std::shared_ptr<ragdoll::Scene> tl);
 	void LoadAndCreateModel(const std::string& fileName);
 private:
 };
